+++ src/main/java/nextstep/subway/domain/line/domain/LineRepository.java
@@ -1,0 +1,7 @@
+package nextstep.subway.domain.line.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+
+}
