+++ src/main/java/nextstep/subway/domain/line/dto/LineRequest.java
@@ -1,0 +1,51 @@
+package nextstep.subway.domain.line.dto;
+
+import nextstep.subway.domain.line.domain.Line;
+
+public class LineRequest {
+    private String name;
+    private String color;
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
+
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public LineRequest(final String name, final String color, final Long upStationId, final Long downStationId, final int distance) {
+        this.name = name;
+        this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+
+    public Line toLine() {
+        return new Line(name, color);
+    }
+}
