+++ src/main/java/nextstep/subway/domain/line/dto/LineResponse.java
@@ -1,0 +1,101 @@
+package nextstep.subway.domain.line.dto;
+
+import nextstep.subway.domain.line.domain.Line;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private List<LineStation> stations;
+    private LocalDateTime createdDate;
+    private LocalDateTime modifiedDate;
+
+    public LineResponse() {
+    }
+
+    public LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.createdDate = createdDate;
+        this.modifiedDate = modifiedDate;
+    }
+
+    public LineResponse(final Long id, final String name, final String color, final List<LineStation> stations, final LocalDateTime createdDate, final LocalDateTime modifiedDate) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.stations = stations;
+        this.createdDate = createdDate;
+        this.modifiedDate = modifiedDate;
+    }
+
+    public static LineResponse of(Line line) {
+        return new LineResponse(line.getId(), line.getName(), line.getColor(), line.getCreatedDate(), line.getModifiedDate());
+    }
+
+    public static LineResponse of(Line line, List<LineStation> lineStations) {
+        return new LineResponse(line.getId(), line.getName(), line.getColor(), lineStations, line.getCreatedDate(), line.getModifiedDate());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public LocalDateTime getModifiedDate() {
+        return modifiedDate;
+    }
+
+    public List<LineStation> getStations() {
+        return stations;
+    }
+
+    public static class LineStation {
+
+        private Long id;
+        private String name;
+        private LocalDateTime createdDate;
+        private LocalDateTime modifiedDate;
+
+        public Long getId() {
+            return id;
+        }
+
+        public String getName() {
+            return name;
+        }
+
+        public LocalDateTime getCreatedDate() {
+            return createdDate;
+        }
+
+        public LocalDateTime getModifiedDate() {
+            return modifiedDate;
+        }
+
+        public LineStation() {
+        }
+
+        public LineStation(final Long id, final String name, final LocalDateTime createdDate, final LocalDateTime modifiedDate) {
+            this.id = id;
+            this.name = name;
+            this.createdDate = createdDate;
+            this.modifiedDate = modifiedDate;
+        }
+    }
+}
