+++ src/main/java/nextstep/subway/domain/section/domain/SectionRepository.java
@@ -1,0 +1,6 @@
+package nextstep.subway.domain.section.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SectionRepository extends JpaRepository<Section, Long> {
+}
