+++ src/main/java/nextstep/subway/domain/station/domain/Station.java
@@ -1,0 +1,43 @@
+package nextstep.subway.domain.station.domain;
+
+import nextstep.subway.domain.common.BaseEntity;
+
+import javax.persistence.*;
+import java.util.Objects;
+
+@Entity
+public class Station extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String name;
+
+    public Station() {
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final Station station = (Station) o;
+        return Objects.equals(id, station.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+}
