+++ src/main/java/nextstep/subway/domain/station/domain/StationRepository.java
@@ -1,0 +1,10 @@
+package nextstep.subway.domain.station.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+    @Override
+    List<Station> findAll();
+}
