+++ src/main/java/nextstep/subway/domain/station/dto/StationRequest.java
@@ -1,0 +1,15 @@
+package nextstep.subway.domain.station.dto;
+
+import nextstep.subway.domain.station.domain.Station;
+
+public class StationRequest {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public Station toStation() {
+        return new Station(name);
+    }
+}
