+++ src/main/java/nextstep/subway/domain/station/dto/StationResponse.java
@@ -1,0 +1,42 @@
+package nextstep.subway.domain.station.dto;
+
+import nextstep.subway.domain.station.domain.Station;
+
+import java.time.LocalDateTime;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+    private LocalDateTime createdDate;
+    private LocalDateTime modifiedDate;
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(), station.getModifiedDate());
+    }
+
+    public StationResponse() {
+    }
+
+    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        this.id = id;
+        this.name = name;
+        this.createdDate = createdDate;
+        this.modifiedDate = modifiedDate;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public LocalDateTime getModifiedDate() {
+        return modifiedDate;
+    }
+}
