--- src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,56 +1,0 @@
-package nextstep.subway.line.domain;
-
-import nextstep.subway.common.BaseEntity;
-import nextstep.subway.section.domain.Section;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-public class Line extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", unique = true)
-    private String name;
-
-    @Column(name = "color")
-    private String color;
-
-    @Embedded
-    private Sections sections = new Sections();
-
-    protected Line() {
-    }
-
-    public Line(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public void update(Line line) {
-        this.name = line.getName();
-        this.color = line.getColor();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public void addSection(final Section section) {
-        this.sections.addSection(section);
-    }
-
-    public List<Section> getSections() {
-        return sections.getStationInOrder();
-    }
-}
