--- src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,7 +1,0 @@
-package nextstep.subway.line.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LineRepository extends JpaRepository<Line, Long> {
-
-}
