--- src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,48 +1,0 @@
-package nextstep.subway.line.ui;
-
-import nextstep.subway.line.application.LineService;
-import nextstep.subway.line.dto.LineRequest;
-import nextstep.subway.line.dto.LineResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/lines")
-public class LineController {
-    private final LineService lineService;
-
-    public LineController(final LineService lineService) {
-        this.lineService = lineService;
-    }
-
-    @PostMapping
-    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
-        LineResponse line = lineService.saveLine(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<LineResponse>> findAllLines() {
-        return ResponseEntity.ok().body(lineService.findAllLine());
-    }
-
-    @PutMapping("/{lineId}")
-    public ResponseEntity<LineResponse> updateLine(@PathVariable("lineId") Long lineId,
-                                                   @RequestBody LineRequest lineRequest) {
-        return ResponseEntity.ok().body(lineService.modifyLine(lineId, lineRequest));
-    }
-
-    @DeleteMapping("/{lineId}")
-    public ResponseEntity<Void> deleteLine(@PathVariable("lineId") Long lineId) {
-        lineService.deleteLine(lineId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/{lineNo}")
-    public ResponseEntity<LineResponse> findLine(@PathVariable("lineNo") Long lineNo) {
-        return ResponseEntity.ok(lineService.findDetailLine(lineNo));
-    }
-}
