--- src/main/java/nextstep/subway/section/domain/SectionRepository.java
@@ -1,6 +1,0 @@
-package nextstep.subway.section.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-}
